+++ gestion_base_adresse/install/sql/upgrade/upgrade_to_1.2.0.sql
@@ -1,28 +1,30 @@
 BEGIN;
 
-CREATE FUNCTION adresse.reset_referencer_com() RETURNS void
+CREATE FUNCTION adresse.add_referencer_com() RETURNS void
     LANGUAGE plpgsql
     AS $$
 BEGIN
 
 DROP INDEX IF EXISTS adresse.idx_tmp_centroide_communes;
 CREATE INDEX idx_tmp_centroide_communes ON adresse.commune_deleguee (ST_PointOnSurface(geom));
-TRUNCATE adresse.referencer_com RESTART IDENTITY;
 
 INSERT INTO adresse.referencer_com
 SELECT
 	c.id_com AS id_com,
-	cd.id_com_del AS id_com_del
+	cd.id_com_del AS id_com_del,
+	'Ajout automatique en fonction des tables commune_deleguee et commune' AS action,
+	NOW() AS date
 FROM adresse.commune_deleguee cd
 JOIN adresse.commune AS c
 	ON St_Contains(c.geom, ST_PointOnSurface(cd.geom))
+ON CONFLICT ON CONSTRAINT referencer_com_pkey DO NOTHING
 ;
 
 DROP INDEX IF EXISTS adresse.idx_tmp_centroide_communes;
 END;
 $$;
 
-COMMENT ON FUNCTION adresse.reset_referencer_com IS 'Permet la réinitialisation de la table adresse.referencer_com en faisant une intersection entre adresse.commune_deleguee et adresse.commune';
+COMMENT ON FUNCTION adresse.add_referencer_com IS 'Permet la réinitialisation de la table adresse.referencer_com en faisant une intersection entre adresse.commune_deleguee et adresse.commune';
 
 CREATE OR REPLACE VIEW adresse.v_export_bal AS
  SELECT ''::text AS uid_adresse,